Create /usr/lib/python3/dist-packages if missing

+++ clean.sh
@@ -3,8 +3,12 @@ PROG_DIR=$(readlink -e $(dirname $0))
 source ${PROG_DIR}/config.sh || exit 1
 
 oldpwd=$(pwd)
-cd usr/lib/python3/dist-packages || exit 1
-rm -rf ${PACKAGE}
-cd $oldpwd
+if [ -d "$DIST_PKGS_DIR" ]; then
+    cd "$DIST_PKGS_DIR"
+    rm -rf ${PACKAGE}
+    cd "$oldpwd"
+fi
 
-rm -fv control.tar.gz data.tar.gz ${PACKAGE}_*_${ARCH}.deb debian-binary
+cd "${PROG_DIR}"
+rm -fv debian-binary control.tar.gz data.tar.gz DEBIAN/md5sums ${PACKAGE}_*_${ARCH}.deb
+cd "$oldpwd"
